Supress SA1402 rule in files with generic and non-generic classes.
